requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
google-generativeai>=0.8.0
Pillow>=10.0.0
